Upstream merge

+++ src/uas/UAS.cc
@@ -997,11 +997,11 @@ void UAS::setExternalControlSetpoint(float roll, float pitch, float yaw, float t
         } else if (joystickMode == Vehicle::JoystickModeRC) {
 
             // Save the new manual control inputs
-            manualRollAngle = roll;
-            manualPitchAngle = pitch;
-            manualYawAngle = yaw;
-            manualThrust = thrust;
-            manualButtons = buttons;
+            manualRollAngle     = roll;
+            manualPitchAngle    = pitch;
+            manualYawAngle      = yaw;
+            manualThrust        = thrust;
+            manualButtons       = buttons;
 
             // Store scaling values for all 3 axes
             const float axesScaling = 1.0 * 1000.0;
@@ -1009,19 +1009,22 @@ void UAS::setExternalControlSetpoint(float roll, float pitch, float yaw, float t
             // Calculate the new commands for roll, pitch, yaw, and thrust
             const float newRollCommand = roll * axesScaling;
             // negate pitch value because pitch is negative for pitching forward but mavlink message argument is positive for forward
-            const float newPitchCommand = -pitch * axesScaling;
-            const float newYawCommand = yaw * axesScaling;
+            const float newPitchCommand  = -pitch * axesScaling;
+            const float newYawCommand    = yaw * axesScaling;
             const float newThrustCommand = thrust * axesScaling;
 
-            //qDebug() << newRollCommand << newPitchCommand << newYawCommand << newThrustCommand;
-
             // Send the MANUAL_COMMAND message
-            mavlink_msg_manual_control_pack_chan(mavlink->getSystemId(),
-                                                 mavlink->getComponentId(),
-                                                 _vehicle->priorityLink()->mavlinkChannel(),
-                                                 &message,
-                                                 this->uasId,
-                                                 newPitchCommand, newRollCommand, newThrustCommand, newYawCommand, buttons);
+            mavlink_msg_manual_control_pack_chan(
+                static_cast<uint8_t>(mavlink->getSystemId()),
+                static_cast<uint8_t>(mavlink->getComponentId()),
+                _vehicle->priorityLink()->mavlinkChannel(),
+                &message,
+                static_cast<uint8_t>(this->uasId),
+                static_cast<int16_t>(newPitchCommand),
+                static_cast<int16_t>(newRollCommand),
+                static_cast<int16_t>(newThrustCommand),
+                static_cast<int16_t>(newYawCommand),
+                buttons);
         }
 
         _vehicle->sendMessageOnLink(_vehicle->priorityLink(), message);
